Makes the test directory into a package.
This allows tests to be run with

  python -m test.test_asyncio <port>

without running into conflicts with Python's built-in unit tests which also have a test.test_asyncio with runs the Python Standard Library tests.
